requirements.txt
Django>=5.2.4
gunicorn
psycopg2-binary
djangorestframework
